+++ source/Private/Convert-ShouldBeFalse.ps1
@@ -176,7 +176,12 @@ function Convert-ShouldBeFalse
         {
             $originalParameterName = $parameterNames.$currentParameter
 
-            $newExtentText += ' -{0} {1}' -f $currentParameter, $commandParameters.$originalParameterName.ExtentText
+            $newExtentText += ' -{0}' -f $currentParameter
+
+            if ($commandParameters.$originalParameterName.ExtentText)
+            {
+                $newExtentText += ' {0}' -f $commandParameters.$originalParameterName.ExtentText
+            }
         }
     }