+++ CHANGELOG.md
@@ -51,3 +51,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Some code cleanup.
 - Fix incorrect position value in Get-PesterCommandParameter.
 - Update PowerShell version requirement to 7.1.
+- `Get-PesterCommandParameter`
+  - Now supports Switch parameters.